pyproject.toml
[project]
name = "string-art"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "imageio>=2.37.0",
    "matplotlib>=3.10.3",
    "numpy>=2.3.1",
    "pyyaml>=6.0.2",
    "scikit-image>=0.25.2",
    "scipy>=1.16.0",
    "torch>=2.7.1",
    "torchvision>=0.22.1",
    "tqdm>=4.67.1",
]
[tool.uv]
package = true
